+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/MiniApp.kt
@@ -111,7 +111,7 @@ abstract class MiniApp internal constructor() {
             defaultConfig = miniAppSdkConfig
             val apiClient = ApiClient(
                 baseUrl = miniAppSdkConfig.baseUrl,
-                rasProjectId = miniAppSdkConfig.rasProjectId,
+                rasProjectOrAppId = miniAppSdkConfig.projectOrAppId,
                 subscriptionKey = miniAppSdkConfig.subscriptionKey,
                 hostAppVersionId = miniAppSdkConfig.hostAppVersionId,
                 isTestMode = miniAppSdkConfig.isTestMode