+++ testapp/src/main/java/com/rakuten/tech/mobile/testapp/ui/settings/AppSettings.kt
@@ -90,6 +90,7 @@ class AppSettings private constructor(context: Context) {
         get() = MiniAppSdkConfig(
             baseUrl = baseUrl,
             rasProjectId = projectId,
+            rasAppId = projectId, // rasAppId is deprecated
             subscriptionKey = subscriptionKey,
             hostAppVersionId = hostAppVersionId,
             // no update for hostAppUserAgentInfo because SDK does not allow changing it at runtime